fully responsive portfolio finished.
There is a bug in the navbar section. The nav-text doesn't want to change to red when I hover it. It used to work but now it doesn't while I didn't change anything. Most likely because of a conflicting code but I haven't found it yet. But I'm happy with the result for now.